+++ integration/VODE/vode_type_true_sdc.F90
@@ -1,7 +1,0 @@
-module vode_type_module
-
-  use amrex_fort_module, only: rt => amrex_real
-
-  implicit none
-
-end module vode_type_module