+++ unit_test/test_jac/main.cpp
@@ -22,6 +22,7 @@ using namespace amrex;
 #include <variables.H>
 #include <unit_test.H>
 #include <react_util.H>
+#include <unit_test_F.H>
 
 int main (int argc, char* argv[])
 {
@@ -114,9 +115,6 @@ void main_main ()
       probin_file_name[i] = probin_file[i];
 
     init_unit_test(probin_file_name.dataPtr(), &probin_file_length);
-
-    // Copy extern parameters from Fortran to C++
-    init_extern_parameters();
 
     // C++ EOS initialization (must be done after Fortran eos_init and init_extern_parameters)
     eos_init(small_temp, small_dens);