+++ CHANGES.md
@@ -1,3 +1,15 @@
+# 21.02
+
+   * Fortran support for the VODE integrator has been removed (#538)
+
+   * Runtime parameters can now be set in the inputs file instead of
+     the probin file (and then they are read in by C++ ParmParse).  If
+     a parameter is set in both places, then the inputs value is used.
+     (#505)
+
+   * Fortran support for simplified-SDC in the VODE integrator has
+     been removed. (#492)
+
 # 21.01
 
    * Microphysics now requires C++17 (gcc >= 7, CUDA >= 11). (#485)